+++ react/__mocks__/intersectionObserverMock.js
@@ -1,0 +1,13 @@
+global.IntersectionObserver = class IntersectionObserver {
+  observe() {
+    return null
+  }
+
+  disconnect() {
+    return null
+  }
+
+  unobserve() {
+    return null
+  }
+}