+++ react/__tests__/OrderBy.test.js
@@ -1,23 +1,31 @@
+/* eslint-disable jest/no-mocks-import */
 /* eslint-env jest */
 import React from 'react'
 import { render, fireEvent } from '@vtex/test-tools/react'
+import { useDevice } from 'vtex.device-detector'
+import { setMobileState } from 'vtex.render-runtime'
 
 import OrderBy from '../OrderBy'
-import { setMobileState } from 'vtex.render-runtime'
 import { useRuntime } from '../__mocks__/vtex.render-runtime'
+
 const mockUseRuntime = useRuntime
 
 const mockSetQuery = jest.fn()
+
 beforeEach(() => {
   jest.clearAllMocks()
 
   mockUseRuntime.mockImplementation(() => ({
     setQuery: mockSetQuery,
   }))
+
+  useDevice.mockImplementation(() => ({
+    isMobile: false,
+  }))
 })
 
 describe('<OrderBy />', () => {
-  const renderComponent = customProps => mobile => {
+  const renderComponent = (customProps) => (mobile) => {
     setMobileState(mobile)
 
     const props = {
@@ -28,6 +36,10 @@ describe('<OrderBy />', () => {
 
     return render(<OrderBy {...props} />)
   }
+
+  afterEach(() => {
+    setMobileState(false)
+  })
 
   it('should shown dropdown box on mobile mode', () => {
     const { container } = renderComponent()(true)
@@ -56,9 +68,5 @@ describe('<OrderBy />', () => {
     const { asFragment } = renderComponent()(true)
 
     expect(asFragment()).toMatchSnapshot()
-  })
-
-  afterEach(() => {
-    setMobileState(false)
   })
 })