+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.98.3] - 2021-04-15
+### Fixed
+- Add new category mapping compatibility to `SearchTitle`
+
+## [3.98.2] - 2021-04-06
+### Added
+- Main tests to the `FilterNavigator`.
+
 ## [3.98.1] - 2021-03-26
 
 ### Fixed